+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## v0.2.1
+
+* Fix `saveAll` on repository method
+* Add `excludeLargeProperties` on `entityOptions` when decorate an `@entity`
+
 ## v0.2.0
 
 * Update [Wiki](https://github.com/pflima92/google-cloud-datastore-inversify/wiki)